+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix bug where rendering manifests results in files being overwritten by subsequent resources with the same kind and name, but different namespace (https://github.com/pulumi/pulumi-kubernetes/pull/1429)
 - Update pulumi dependency to fix python Resource.get() functions (https://github.com/pulumi/pulumi-kubernetes/pull/1480)
 - Upgrade to Go1.16 (https://github.com/pulumi/pulumi-kubernetes/pull/1486)
+- Fix bug preventing helm chart being located in bitnami repo (https://github.com/pulumi/pulumi-kubernetes/pull/1491)
 
 ## 2.8.2 (February 23, 2021)